+++ mjj_mall/routes/companyRoute.js
@@ -1,8 +1,8 @@
 const express = require('express');
 const router = express.Router();
 
-const controller = require('./companyRouteController');
-const validation = require('./companyValidation');
+const controller = require('../middlewares/company/companySignUp');
+const validation = require('../middlewares/company/companyValidation');
 
 router.post('/sign-up',validation.companyValidation, controller.signUp);
 router.post('/test-url',controller.testAndroidGet);
